Allow configuration of network wait time.
Add a --networkwait CLI argument and a NetworkWaitSeconds preferences
key to allow configuration of the network wait delay in
checkandinstallatstartup mode.

This allows administrators to delay for longer if the network takes a
long time to come up, or to skip the network wait by setting the wait to
zero.